+++ lib/css-rewrite.rb
@@ -1,8 +1,6 @@
 module CssRewrite
-  autoload :CallRewriter,  'css-rewrite/call_rewriter'
   autoload :Config,        'css-rewrite/config'
   autoload :Postprocessor, 'css-rewrite/postprocessor'
-  autoload :RegexRewriter, 'css-rewrite/regex_rewriter'
   autoload :Rewriter,      'css-rewrite/rewriter'
 
   def self.configure