+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.0.0
+===
+- Changed interface significantly to avoid parsing files that don't need to be
+  processed. Now, pass a filename or regex to `config.rewrite` to indicate the
+  files that should be handled by the rewriter.
+
 1.0.0
 ===
 - Birthday!